refactor: move user service to service/user
