+++ internal/service/user/service.go
@@ -1,13 +1,14 @@
-package service
+package user
 
 import (
 	"fmt"
 
 	"github.com/nix-united/golang-gin-boilerplate/internal/model"
 	"github.com/nix-united/golang-gin-boilerplate/internal/request"
+	"github.com/nix-united/golang-gin-boilerplate/internal/service"
 )
 
-//go:generate mockgen -source=$GOFILE -destination=user_service_mock_test.go -package=${GOPACKAGE}_test -typed=true
+//go:generate mockgen -source=$GOFILE -destination=service_mock_test.go -package=${GOPACKAGE}_test -typed=true
 
 type userRepository interface {
 	FindUserByEmail(email string) (model.User, error)
@@ -18,14 +19,14 @@ type encryptor interface {
 	Encrypt(str string) (string, error)
 }
 
-// UserService provides a use case level for the user entity
-type UserService struct {
+// Service provides a use case level for the user entity
+type Service struct {
 	userRepository userRepository
 	encryptor      encryptor
 }
 
-func NewUserService(userRepository userRepository, enencryptor encryptor) UserService {
-	return UserService{
+func NewService(userRepository userRepository, enencryptor encryptor) Service {
+	return Service{
 		userRepository: userRepository,
 		encryptor:      enencryptor,
 	}
@@ -34,14 +35,14 @@ func NewUserService(userRepository userRepository, enencryptor encryptor) UserSe
 // CreateUser Create takes a request with new user credentials and registers it.
 // An error will be returned if a user exists in the system, or
 // if an error occurs during interaction with the database.
-func (s UserService) CreateUser(req request.RegisterRequest) error {
+func (s Service) CreateUser(req request.RegisterRequest) error {
 	user, err := s.userRepository.FindUserByEmail(req.Email)
 	if err != nil {
 		return fmt.Errorf("find user by email: %w", err)
 	}
 
 	if user.ID != 0 {
-		return NewErrUserAlreadyExists(
+		return service.NewErrUserAlreadyExists(
 			"user already exist",
 			"store a user",
 		)
