+++ internal/service/user/service_mock_test.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=user_service.go -destination=user_service_mock_test.go -package=service_test -typed=true
+//	mockgen -source=user_service.go -destination=service_mock_test.go -package=user_test -typed=true
 //
 
-// Package service_test is a generated GoMock package.
-package service_test
+// Package user_test is a generated GoMock package.
+package user_test
 
 import (
 	reflect "reflect"
